(memcache): Implement GAT command
The GAT command is implemented for the memcache protocol. The command
gets one or more keys and sets an expiry on the keys.

The command structure reuses much of the mget command code path,
differing where the expiry of the keys is set.

The command is registered as hidden as it has checks for the protocol to
be memcache only.

Signed-off-by: Abhijat Malviya <abhijat@dragonflydb.io>