chore: add dynamic stash buckets
Before: our Dash segments contained both regular and stash buckets.
Each segment was allocated in one shot.
The the segment parameters was carefully chosen to be "allocator friendly" so that
segment size was very close to what mimalloc actually allocates a block with size close to
the segment size. This indeed has a good memory locality.

However when a segment is split, the occupancy rate of two new segments drops to 47-50%
which causes memory spikes for small value workloads.

This change alters the structure of the segment. Now its stash buckets will be allocated separately
and lazily when an item can not be inserted into its home buckets. When a segment is split,
its stash buckets will be most likely deallocated. So for example, a 56+4 bucket segment will be split into
two 56 bucket segments, so the occupancy rate of each one of them will at ~53.5% percent.
Moreover, we will be able to use more stash buckets in the future to be able reach near 100% occupancy rate
because now the stash bucket count can be chosen separately without affecting the block size of the "stripped" segment.

Signed-off-by: Roman Gershman <roman@dragonflydb.io>